compatibility with spring tests

+++ sdk/spring/scripts/compatibility_delete_version.py
@@ -14,7 +14,10 @@ import time
 import argparse
 
 from log import log
-from _constants import get_spring_boot_version_tag_prefix, SPRING_BOOT_MAJOR_3_VERSION_TAG_PREFIX
+from _constants import (get_spring_boot_version_tag_prefix,
+                        SPRING_BOOT_MAJOR_3_VERSION_TAG_PREFIX,
+                        should_skip_artifacts_when_adding_dependency_management,
+                        should_skip_artifacts_when_adding_dependency_management_with_spring_version)
 
 
 IGNORED_ARTIFACTS = {'com.github.tomakehurst:wiremock-jre8'}
@@ -44,7 +47,7 @@ def main():
     args = get_args()
     version_tag_prefix = get_spring_boot_version_tag_prefix(args.spring_boot_dependencies_version)
     ignored_artifacts = get_ignored_artifacts(args.spring_boot_dependencies_version)
-    find_all_poms_do_version_control("./sdk/spring", version_tag_prefix, ignored_artifacts)
+    find_all_poms_do_version_control("./sdk/spring", args.spring_boot_dependencies_version, version_tag_prefix, ignored_artifacts)
     elapsed_time = time.time() - start_time
     log.info('elapsed_time = {}'.format(elapsed_time))
 
@@ -54,12 +57,16 @@ def change_to_repo_root_dir():
     os.chdir('../../..')
 
 
-def find_all_poms_do_version_control(directory, version_tag_prefix, ignored_artifacts):
+def find_all_poms_do_version_control(directory, spring_boot_version, version_tag_prefix, ignored_artifacts):
     external_dependencies_set = external_dependencies_managed_list(version_tag_prefix)
     for root, dirs, files in os.walk(directory):
         for file_name in files:
             if file_name.startswith('pom') and file_name.endswith('.xml'):
                 file_path = join(root, file_name)
+                if (should_skip_artifacts_when_adding_dependency_management_with_spring_version(spring_boot_version, file_path)
+                    or should_skip_artifacts_when_adding_dependency_management(file_path)):
+                    log.warn("Skip deleting version for file: " + file_path)
+                    continue
                 delete_dependency_version(file_path, ignored_artifacts, external_dependencies_set)
